# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Community Hero – Intelligent Issue Resolution API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List complaints
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: priority
          schema: { type: string }
        - in: query
          name: departmentId
          schema: { type: integer }
        - in: query
          name: publicAssetId
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of complaints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintList"
    post:
      operationId: submitComplaint
      tags: [complaints]
      summary: Submit a new complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Complaint submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintDetail"
  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get complaint by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Complaint detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /complaints/ticket/{ticketId}:
    get:
      operationId: getComplaintByTicket
      tags: [complaints]
      summary: Get complaint by ticket ID
      parameters:
        - in: path
          name: ticketId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Complaint detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /complaints/{id}/analyze:
    post:
      operationId: analyzeComplaint
      tags: [complaints]
      summary: Run AI analysis on a complaint
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysisResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /complaints/{id}/assign:
    post:
      operationId: assignWorker
      tags: [complaints]
      summary: Assign nearest qualified worker
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /complaints/{id}/verify:
    post:
      operationId: verifyCompletion
      tags: [complaints]
      summary: AI verification of completion
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
  /complaints/{id}/citizen-feedback:
    post:
      operationId: submitCitizenFeedback
      tags: [complaints]
      summary: Citizen rates and verifies resolution
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CitizenFeedbackInput"
      responses:
        "200":
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
      required: [success]
    ComplaintInput:
      type: object
      properties:
        publicAssetId: { type: ["integer", "null"] }
        citizenName: { type: ["string", "null"] }
        citizenPhone: { type: ["string", "null"] }
        citizenEmail: { type: ["string", "null"] }
        isAnonymous: { type: boolean }
        description: { type: ["string", "null"] }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        address: { type: ["string", "null"] }
        ward: { type: ["string", "null"] }
        imageBase64: { type: ["string", "null"] }
    ComplaintDetail:
      type: object
      required: [id, ticketId, status, createdAt]
      properties:
        id: { type: integer }
        ticketId: { type: string }
        publicAssetId: { type: ["integer", "null"] }
        citizenName: { type: ["string", "null"] }
        citizenPhone: { type: ["string", "null"] }
        isAnonymous: { type: boolean }
        description: { type: ["string", "null"] }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        address: { type: ["string", "null"] }
        ward: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: ["string", "null"] }
        priorityScore: { type: ["number", "null"] }
        departmentId: { type: ["integer", "null"] }
        assignedWorkerId: { type: ["integer", "null"] }
        estimatedResolutionHours: { type: ["number", "null"] }
        resolvedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ComplaintList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ComplaintDetail"
        total: { type: integer }
    AiAnalysisResult:
      type: object
      required: [complaintId, issueCategory]
      properties:
        complaintId: { type: integer }
        issueCategory: { type: ["string", "null"] }
        issueDescription: { type: ["string", "null"] }
        detectedObjects:
          type: array
          items: { type: string }
        suggestedDepartment: { type: ["string", "null"] }
        severity: { type: ["string", "null"] }
        confidence: { type: ["number", "null"] }
        estimatedResolutionHours: { type: ["number", "null"] }
        equipmentRequired:
          type: array
          items: { type: string }
        workersRequired: { type: ["integer", "null"] }
        requiredSkills:
          type: array
          items: { type: string }
        emergencyLevel: { type: ["string", "null"] }
    AssignmentResult:
      type: object
      required: [assignmentId, workerId, workerName]
      properties:
        assignmentId: { type: integer }
        workerId: { type: integer }
        workerName: { type: string }
        workerPhone: { type: string }
        distanceKm: { type: ["number", "null"] }
        estimatedArrivalMinutes: { type: ["number", "null"] }
    VerificationInput:
      type: object
      properties:
        workerNotes: { type: ["string", "null"] }
        beforeImageBase64: { type: ["string", "null"] }
        afterImageBase64: { type: ["string", "null"] }
    VerificationResult:
      type: object
      required: [verdict, confidence]
      properties:
        verdict: { type: string }
        confidence: { type: number }
        aiNotes: { type: ["string", "null"] }
        isEscalated: { type: boolean }
        escalationReason: { type: ["string", "null"] }
    CitizenFeedbackInput:
      type: object
      properties:
        citizenVerified: { type: boolean }
        citizenRating: { type: ["integer", "null"] }
        citizenComment: { type: ["string", "null"] }
...